pyproject.toml
[tool.poetry]
name = "bty-chtbt"
version = "0.1.0"
description = ""
authors = ["nusenyaw66 <nusenyaw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
langchain-core = ">=0.2.0,<0.4.0"
langchain = "^0.3.0"
langchain-community = "^0.3.22"
langchain-chroma = "^0.2.3"
langchain-huggingface = "^0.1.0"
langchain-text-splitters = "^0.3.8"
transformers = "^4.44.0"
faiss-cpu = "^1.11.0"
pandas = "^2.2.0"
pinecone = "^6.0"
openai = "^1.7.0"
tiktoken = "^0.9.0"
flask = "^3.1.0"
flask-cors = "^5.0.0"
line-bot-sdk = "3.17.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
